Take out git option
It seems like the behavior on this can be inconsistent (adds git, even if you uncheck the box, if you have another project with git open). Taking it out for now, but may want to check later and see if that issue's been resolved and, if so, add it back in to our template. Works fine if you don't have another R project open when you open the new project.